LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/reporting-subresource-corp.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt due to access control checks.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not a constructor (evaluating 'new global.ReportingObserver')
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/text-plain.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/text-plain.txt due to access control checks.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not a constructor (evaluating 'new global.ReportingObserver')
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/text-plain.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/text-plain.txt due to access control checks.
CONSOLE MESSAGE: ReferenceError: Can't find variable: ReportingObserver
CONSOLE MESSAGE: ReferenceError: Can't find variable: ReportingObserver
CONSOLE MESSAGE: ReferenceError: Can't find variable: ReportingObserver
CONSOLE MESSAGE: ReferenceError: Can't find variable: ReportingObserver

Harness Error (FAIL), message = ReferenceError: Can't find variable: ReportingObserver

PASS [document] same-origin
PASS [document] blocked by CORP: same-origin
FAIL [document] blocked due to COEP promise_test: Unhandled rejection with value: object "TypeError: undefined is not a constructor (evaluating 'new global.ReportingObserver')"
FAIL [document] blocked during redirect promise_test: Unhandled rejection with value: object "TypeError: undefined is not a constructor (evaluating 'new global.ReportingObserver')"
FAIL [dedicated worker] same-origin assert_unreached: Worker.onerror Reached unreachable code
FAIL [dedicated worker] blocked by CORP: same-origin assert_unreached: Worker.onerror Reached unreachable code
FAIL [dedicated worker] blocked due to COEP assert_unreached: Worker.onerror Reached unreachable code
FAIL [dedicated worker] blocked during redirect assert_unreached: Worker.onerror Reached unreachable code
FAIL [shared worker] same-origin assert_unreached: Worker.onerror Reached unreachable code
FAIL [shared worker] blocked by CORP: same-origin assert_unreached: Worker.onerror Reached unreachable code
FAIL [shared worker] blocked due to COEP assert_unreached: Worker.onerror Reached unreachable code
FAIL [shared worker] blocked during redirect assert_unreached: Worker.onerror Reached unreachable code
TIMEOUT [service worker] same-origin Test timed out
NOTRUN [service worker] blocked by CORP: same-origin
NOTRUN [service worker] blocked due to COEP
NOTRUN [service worker] blocked during redirect
NOTRUN [between service worker and page] same-origin
NOTRUN [between service worker and page] blocked by CORP: same-origin
NOTRUN [between service worker and page] blocked due to COEP
NOTRUN [between service worker and page] blocked during redirect
NOTRUN destination: script
